requirements.txt
pymysql==1.1.0
pytest==7.4.3
behave==1.2.6
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.4.2
alembic==1.12.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6 